requirements.txt
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
Pillow==9.4.0
psycopg2-binary==2.9.5
pytest==7.2.2
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dotenv==1.0.0